feat: add support for null / nan
