+++ packages/core-util/src/sdks/number.ts
@@ -1,4 +1,4 @@
-import type { Primitive } from '@iot-app-kit/core';
+import type { Primitive } from '@iot-app-kit/helpers';
 
 /**
  * Rounds a number to a given precision
