+++ packages/dashboard/src/components/csvDownloadButton/types.ts
@@ -1,4 +1,4 @@
-import { type Primitive } from '@iot-app-kit/core';
+import type { Primitive } from '@iot-app-kit/helpers';
 import { type Quality } from '@aws-sdk/client-iotsitewise';
 
 export type CSVDownloadObject = {
