+++ packages/data-mocked/src/iot-sitewise/handlers/batchGetAssetPropertyValue/batchGetAssetPropertyValueHandler.ts
@@ -67,8 +67,9 @@ export class AssetPropertyValueFactory {
       ...this.#createDefaults(),
       ...partialAssetPropertyValue,
       value: {
+        // nullValue: { valueType: 'I' },
         integerValue: faker.number.int({ min: 0, max: 100 }),
-      },
+      } as any,
     };
 
     return assetPropertyValue;
@@ -81,8 +82,9 @@ export class AssetPropertyValueFactory {
       ...this.#createDefaults(),
       ...partialAssetPropertyValue,
       value: {
+        // nullValue: { valueType: 'D' },
         doubleValue: faker.number.float({ min: 0, max: 100, precision: 0.001 }),
-      },
+      } as any,
     };
 
     return assetPropertyValue;
@@ -95,8 +97,9 @@ export class AssetPropertyValueFactory {
       ...this.#createDefaults(),
       ...partialAssetPropertyValue,
       value: {
+        // nullValue: { valueType: 'B' },
         booleanValue: faker.datatype.boolean(),
-      },
+      } as any,
     };
 
     return assetPropertyValue;
@@ -109,8 +112,9 @@ export class AssetPropertyValueFactory {
       ...this.#createDefaults(),
       ...partialAssetPropertyValue,
       value: {
+        // nullValue: { valueType: 'S' },
         stringValue: faker.helpers.arrayElement(['ON', 'OFF']),
-      },
+      } as any,
     };
 
     return assetPropertyValue;
@@ -206,6 +210,7 @@ export function batchGetAssetPropertyValueHistoryHandler() {
     };
 
     await delay();
+
     return HttpResponse.json(response, { status: 200 });
   });
 }
