+++ packages/react-components/src/components/gauge/converters/convertDataset.ts
@@ -1,4 +1,4 @@
-import { type Primitive } from '@iot-app-kit/core';
+import type { Primitive } from '@iot-app-kit/helpers';
 
 export const convertDataset = (gaugeValue?: Primitive) => {
   return {
