+++ packages/react-components/src/components/resource-explorers/types/resources.ts
@@ -1,4 +1,5 @@
 import type { AssetSummary } from '@aws-sdk/client-iotsitewise';
+import { type Primitive } from '@iot-app-kit/helpers';
 
 export interface AssetModelResource {
   assetModelId: string;
@@ -59,6 +60,6 @@ export type TimeSeriesResourceWithLatestValue =
 
 export type DataStreamResourceWithLatestValue<DataStreamResource> =
   DataStreamResource & {
-    latestValue?: number | string | boolean;
+    latestValue?: Primitive;
     latestValueTimestamp?: number;
   };
