+++ packages/react-components/src/queries/useAssetPropertyValues/utils/toDataPoint.ts
@@ -1,5 +1,6 @@
 import { NANO_SECOND_IN_MS, SECOND_IN_MS } from './timeConstants';
-import type { DataPoint, Primitive } from '@iot-app-kit/core';
+import type { Primitive } from '@iot-app-kit/helpers';
+import type { DataPoint } from '@iot-app-kit/core';
 import type {
   AssetPropertyValue,
   TimeInNanos,
