+++ packages/react-components/src/utils/getPreciseValue.ts
@@ -1,5 +1,5 @@
-import { type Primitive } from '@iot-app-kit/core';
+import type { Primitive } from '@iot-app-kit/helpers';
 import { isNumeric, round } from '@iot-app-kit/core-util';
 
 export const getPreciseValue = (value: Primitive, significantDigits = 4) =>
-  isNumeric(value) ? `${round(value, significantDigits)}` : value.toString();
+  isNumeric(value) ? `${round(value, significantDigits)}` : value?.toString();
