+++ packages/react-components/src/utils/sort.ts
@@ -1,4 +1,5 @@
-import type { DataPoint, Primitive } from '@iot-app-kit/core';
+import type { Primitive } from '@iot-app-kit/helpers';
+import type { DataPoint } from '@iot-app-kit/core';
 
 /**
  * Sorts points in order of their points values.
