+++ packages/source-iottwinmaker/src/common/types.ts
@@ -1,1 +1,1 @@
-export type { Primitive } from '@iot-app-kit/core';
+export type { Primitive } from '@iot-app-kit/helpers';
